requirements.txt
fpdf2==2.8.5
pyelftools==0.32